Add files via upload
runs FFT on the "Test 1.wav" file which can be found in the google drive, pretty sure I squashed all the bugs but there are probably still some issues. I tried running it using a .wav file with pure sinewaves as opposed to a piano and it was having some trouble so it'll still need some minor tweaks.